move scan serialization tests to kernel tests
