+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -358,23 +358,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "scan_dataset_serialization_test",
-    size = "medium",
-    srcs = ["scan_dataset_serialization_test.py"],
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/experimental/ops:scan_ops",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-    ],
-)
-
-tf_py_test(
     name = "sequence_dataset_serialization_test",
     size = "medium",
     srcs = ["sequence_dataset_serialization_test.py"],
